move has_tappables
