+++ packages/flame/lib/game.dart
@@ -10,7 +10,7 @@ export 'src/game/mixins/fps_counter.dart';
 export 'src/game/mixins/game.dart';
 export 'src/game/mixins/has_draggables.dart';
 export 'src/game/mixins/has_hoverables.dart';
-export 'src/game/mixins/has_tappables.dart';
+export 'src/events/flame_game_mixins/has_tappables.dart';
 export 'src/game/mixins/loadable.dart';
 export 'src/game/mixins/single_game_instance.dart';
 export 'src/game/projector.dart';
